v1.0.11
Fixed displayed stats being removed off of items when a potion coating on it expires
Fixed cauldrons deleting items when a catalyst is clicked on it
Attempted fix to prevent progress rollbacks when using mysql database
Fixed farming luck icon not being displayed (requires language file reset)
Fixed 6x birch planks recipe accepting warped hyphae and nylium as ingredients (requires grid recipe file reset)
Blocks affected by explosions should now also be communicated to plugins like coreprotect instead of being ignored
Fixed transmutation not working properly while worldguard is installed
Fixed 0-damage hits (such as snowballs) never actually hitting
Fixed power attack splash damage radius being noted as a percentile instead of number
Fixed power attacks not working
Fixed damage indicators (without DecentHolograms) not disappearing
Fixed elytra crafting recipe not rewarding smithing EXP (requires grid recipe file reset)
Fixed farming luck not actually applying to harvested crops
Fixed arrows always ignoring armor (whoops lol)
Fixed multishot crossbow arrow duplication
Fixed weapon coating not working for heavy weapons
Fixed additional damage instances causing previous damage hits to be ignored (notably radiant damage from weapon coating)
Fixed potion effect charges of instant custom effects not depleting with coatings
FIXED Issue where upgrading to netherite wasn't applying default netherite stats to said item, its armors not getting knockback resistance
FIXED Farming luck not actually applying to harvested crops
FIXED Issue where soft reset resulted in the player having double the intended skill points
FIXED Using tridents ranged now gives you heavy weapons exp (or light weapons exp if the trident is marked to be light)
TWEAK Every single custom item is now registered to the custom item registry, including some vanilla items with custom attributes
RESOURCE PACK Updated!

1.20.5+ compatibility
Removed most compiler warnings and cleaned up imports

+++ core/src/main/java/me/athlaeos/valhallammo/item/arrow_attributes/implementations/DragonFireballArrow.java
@@ -25,7 +25,7 @@ public class DragonFireballArrow extends ArrowBehavior {
             boolean isIncendiary = ((int) args[1]) == 1;
 
             double multiplier = AccumulativeStatManager.getCachedStats("EXPLOSION_RADIUS_MULTIPLIER", e.getEntity(), 10000, true);
-            radius *= (1 + multiplier);
+            radius *= (float) (1 + multiplier);
 
             DragonFireball dragonFireball = e.getEntity().launchProjectile(DragonFireball.class, e.getProjectile().getVelocity());
             dragonFireball.setShooter(e.getEntity());
