+++ core/src/main/java/me/athlaeos/valhallammo/item/arrow_attributes/implementations/ExplodingArrow.java
@@ -34,7 +34,7 @@ public class ExplodingArrow extends ArrowBehavior {
             Location hit = getHitLocation(e);
             if (e.getEntity().getShooter() instanceof Entity s){
                 double multiplier = AccumulativeStatManager.getCachedStats("EXPLOSION_RADIUS_MULTIPLIER", s, 10000, true);
-                radius *= (1 + multiplier);
+                radius *= (float) (1 + multiplier);
                 e.getEntity().getWorld().createExplosion(hit, radius, fire, destructive, s);
             } else
                 e.getEntity().getWorld().createExplosion(hit, radius, fire, destructive, e.getEntity());
